fix: Build Matrix Verifier race condition, evidence bounds, bloat reduction
Build Matrix Verifier:
- Split runMatches into runMatchesIdentity + full runMatches to handle
  GitHub's in_progress event arriving after a build completes
- Lifecycle check now accepts both in_progress and completed statuses
  for in_progress events, preventing false rejections

Evidence verification:
- Raise MAX_SINGLE_FILE_BYTES from 2 GB to 4 GB for SparkEngineLib.lib
  (monolithic MSVC static lib, 435+ translation units)

Bloat reduction (-1,184 lines net):
- Split TelemetrySpool.cpp (849 lines) into three files: core (332),
  format serialization (252), filesystem I/O (307)
- Extract BC block compression from TextureCompressor.cpp (539->365)
  into TextureBlockCompression.cpp (188)
- Extract crash artifact directory from CrashHandlerSupport.h (367->226)
  into CrashArtifactDirectory.h (167)
- Trim SaveSystemTypes.h (382->186), Components.h (323->291),
  EditorPluginManager.cpp (545->500)
- Add 5 borderline files to .bloat-baseline.txt

Tooling:
- Create tools/check-test-registration.sh (validates every Test*.cpp
  is referenced in Tests/CMakeLists.txt)
