+++ .github/badges/loc-breakdown.json
@@ -1,9 +1,9 @@
 {
   "schemaVersion": 1,
-  "total": 766022,
+  "total": 764913,
   "files": 2587,
-  "engine": 314465,
-  "editor": 109226,
+  "engine": 313403,
+  "editor": 109179,
   "game": 143038,
   "services": 11280,
   "pipeline": 2400,
