+++ .github/badges/loc.json
@@ -1,7 +1,7 @@
 {
   "schemaVersion": 1,
   "label": "C++ lines of code",
-  "message": "766,022",
+  "message": "764,913",
   "color": "blue",
   "logo": "cplusplus",
   "cacheSeconds": 300
