+++ Tools/buildmatrix/verify_external_evidence.py
@@ -37,7 +37,7 @@ MAX_METADATA_BYTES = 1024 * 1024
 MAX_PENDING_RECEIPT_BYTES = 1024 * 1024
 MAX_FILE_COUNT = 100_000
 MAX_TOTAL_BYTES = 8 * 1024 * 1024 * 1024
-MAX_SINGLE_FILE_BYTES = 2 * 1024 * 1024 * 1024
+MAX_SINGLE_FILE_BYTES = 4 * 1024 * 1024 * 1024
 MAX_COMPRESSED_ARTIFACT_BYTES = 4 * 1024 * 1024 * 1024
 
 PROFILE_BUILD_DIRS = {
